Move random rotation generation to new module
